# oapispec/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the NBRGLM Nexeres API Spec documentation.",
        "title": "NBRGLM Nexeres API Spec",
        "termsOfService": "https://nbrglm.com/nexeres/terms",
        "contact": {
            "name": "NBRGLM Support",
            "url": "https://nbrglm.com/support",
            "email": "contact@nbrglm.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "host": "localhost:3360",
    "basePath": "/",
    "paths": {
...
        "/api/auth/refresh": {
            "post": {
                "description": "Handles token refresh requests.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh Token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Refresh token",
                        "name": "X-NEXERES-Refresh-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New tokens",
                        "schema": {
                            "$ref": "#/definitions/handlers.RefreshTokenResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid or missing tokens",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or expired tokens - Proceed to Login",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/signup": {
            "post": {
                "description": "Handles user registration requests.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Signup",
                "parameters": [
                    {
                        "description": "User Signup Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UserSignupData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Signup Result",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserSignupResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid Invite Token or Missing Invite Token or Domain Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.RefreshTokenResult": {
            "type": "object",
            "properties": {
                "tokens": {
                    "$ref": "#/definitions/tokens.Tokens"
                }
            }
        },
...
        "handlers.UserSignupData": {
            "type": "object",
            "required": [
                "confirmPassword",
                "email",
                "firstName",
                "lastName",
                "password"
            ],
            "properties": {
                "confirmPassword": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "inviteToken": {
                    "description": "Optional invite token for signup",
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 8
                }
            }
        },
        "handlers.UserSignupResult": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "debug": {
                    "description": "DebugMessage is a technical message that can be used for debugging.",
                    "type": "string"
                },
                "message": {
                    "description": "Message is a user-friendly message that can be displayed to the end user.",
                    "type": "string"
                }
            }
        },
...
        "tokens.Tokens": {
            "type": "object",
            "properties": {
                "refreshToken": {
                    "description": "RefreshToken is the generated refresh token.\n\nThis is base64.RawURLEncoding encoded.\nDO NOT DECODE IT WHILE RETRIEVING IT FROM THE COOKIES/CLIENT.",
                    "type": "string"
                },
                "refreshTokenExpiry": {
                    "description": "RefreshTokenExpiry is the expiration time of the refresh token.",
                    "type": "string"
                },
                "sessionId": {
                    "description": "SessionId is the unique identifier for the session.",
                    "type": "string"
                },
                "sessionToken": {
                    "description": "SessionToken is the generated session token.\n\nThis is a jwt which is base64.RawURLEncoding encoded.\nYOU NEED TO DECODE IT WHILE RETRIEVING IT FROM THE COOKIES/CLIENT.\nDO NOT USE IT AS IS. VALIDATION WILL FAIL WITHOUT DECODING.\nONLY WHEN DECODED, YOU SHOULD PASS IT TO THE THINGS THAT REQUIRE THE SESSION TOKEN.",
                    "type": "string"
                },
                "sessionTokenExpiry": {
                    "description": "SessionTokenExpiry is the expiration time of the session token.",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "type": "apiKey",
            "name": "X-NEXERES-API-Key",
            "in": "header"
        },
        "AdminHeaderAuth": {
            "type": "apiKey",
            "name": "X-NEXERES-Admin-Token",
            "in": "header"
        },
        "RefreshHeaderAuth": {
            "type": "apiKey",
            "name": "X-NEXERES-Refresh-Token",
            "in": "header"
        },
        "SessionHeaderAuth": {
            "type": "apiKey",
            "name": "X-NEXERES-Session-Token",
            "in": "header"
        }
    },
...
}